quiche_server
Function mint_token

pub(crate) fn mint_token(hdr: &Header<'_>, src: &SocketAddr) -> Vec<u8>

Generate a stateless retry token.

The token includes the static string "quiche" followed by the IP address of the client and by the original destination connection ID generated by the client.

Note that this function is only an example and doesn’t do any cryptographic authenticate of the token. It should not be used in production system.