quiche_server

Function set_txtime_sockopt

pub(crate) fn set_txtime_sockopt(sock: &UdpSocket) -> Result<()>

Set SO_TXTIME socket option.

This socket option is set to send to kernel the outgoing UDP packet transmission time in the sendmsg syscall.

Note that this socket option is set only on linux platforms.